[Result] Rename classes
