+++ src/Psl/Result/Failure.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Psl\Asio;
+namespace Psl\Result;
 
 use Exception;
 use Psl;
@@ -13,9 +13,9 @@ use Psl;
  * @template    T
  * @template    Te of Exception
  *
- * @implements  IResultOrExceptionWrapper<T>
+ * @implements  ResultInterface<T>
  */
-final class WrappedException implements IResultOrExceptionWrapper
+final class Failure implements ResultInterface
 {
     /**
      * @psalm-var Te
